Add files via upload
Overview:
Giving the loan to a costumer is not one step process. Company have to do analysis on many confidential data provided by the customer. So automate the loan eligibility process (real time) based on customer detail provide while filling online application form, in the above data set we have collected all the necessary details. 

We have to automate this process by identifying the customer segments and those who are eligible for loan amount so that they can specifically target these customers.

The aim of behind creating this model is to predict loan eligibility of different categories of people.
For model building I have used the “ORANGE” tool which is an open source data visualization and analysis tool, where data mining is done through visual programming or Python scripting. The tool has components for machine learning

I have also added exploration sheet, which will help you to get more detailed explanation for the data set given above.

With the help of different graphs and histograms we can interpret and understand that which attribute  plays an important role in determining whether the person is
eligible for loan or not and also helps the company to target suitable customers for the loan.